+++ scripts/test-weights.json
@@ -237,7 +237,8 @@
   "test/eval-cross-modal-batch.test.ts": 141,
   "test/eval-export.test.ts": 2399,
   "test/eval-gate.test.ts": 3884,
-  "test/eval-longmemeval.slow.test.ts": 359087,
+  "test/eval-longmemeval-e2e.slow.test.ts": 196000,
+  "test/eval-longmemeval.slow.test.ts": 42000,
   "test/eval-prune.test.ts": 18396,
   "test/eval-replay-gate.test.ts": 34958,
   "test/eval-replay-metadata-skip.test.ts": 3430,
